JavaScript (v3): Move SNS client to libs folder.
