+++ javascriptv3/example_code/sns/libs/snsClient.js
