+++ javascriptv3/example_code/sns/src/sns_confirmsubscription.js
@@ -22,7 +22,7 @@ node sns_confirmsubscription.js
 // snippet-start:[sns.JavaScript.subscriptions.confirmSubscriptionV3]
 // Import required AWS SDK clients and commands for Node.js
 import {ConfirmSubscriptionCommand } from "@aws-sdk/client-sns";
-import {snsClient } from "./libs/snsClient.js";
+import {snsClient } from "../libs/snsClient.js";
 
 // Set the parameters
 const params = {
