+++ javascriptv3/example_code/sns/src/sns_deletetopic.js
@@ -19,7 +19,7 @@ node sns_deletetopic.js
 
 // Import required AWS SDK clients and commands for Node.js
 import {DeleteTopicCommand } from "@aws-sdk/client-sns";
-import {snsClient } from "./libs/snsClient.js";
+import {snsClient } from "../libs/snsClient.js";
 
 // Set the parameters
 const params = { TopicArn: "TOPIC_ARN" }; //TOPIC_ARN
