+++ javascriptv3/example_code/sns/src/sns_getsmstype.js
@@ -17,7 +17,7 @@ node sns_getsmstype.js
 // snippet-start:[sns.JavaScript.SMS.getSMSAttributesV3]
 // Import required AWS SDK clients and commands for Node.js
 import {GetSMSAttributesCommand } from "@aws-sdk/client-sns";
-import {snsClient } from "./libs/snsClient.js";
+import {snsClient } from "../libs/snsClient.js";
 
 // Set the parameters
 var params = {
