+++ javascriptv3/example_code/sns/src/sns_gettopicattributes.js
@@ -17,7 +17,7 @@ node sns_gettopicattributes.js
 // snippet-start:[sns.JavaScript.topics.getTopicAttributesV3]
 // Import required AWS SDK clients and commands for Node.js
 import {GetTopicAttributesCommand } from "@aws-sdk/client-sns";
-import {snsClient } from "./libs/snsClient.js";
+import {snsClient } from "../libs/snsClient.js";
 
 // Set the parameters
 const params = { TopicArn: "TOPIC_ARN" }; // TOPIC_ARN
