+++ javascriptv3/example_code/sns/src/sns_listnumbersoptedout.js
@@ -17,7 +17,7 @@ node sns_listnumbersoptedout.js
 
 // Import required AWS SDK clients and commands for Node.js
 import {ListPhoneNumbersOptedOutCommand } from "@aws-sdk/client-sns";
-import {snsClient } from "./libs/snsClient.js";
+import {snsClient } from "../libs/snsClient.js";
 
 const run = async () => {
   try {
