+++ javascriptv3/example_code/sns/src/sns_publishsms.js
@@ -18,7 +18,7 @@ node sns_publishsms.js
 // snippet-start:[sns.JavaScript.SMS.publishV3]
 // Import required AWS SDK clients and commands for Node.js
 import {PublishCommand } from "@aws-sdk/client-sns";
-import {snsClient } from "./libs/snsClient.js";
+import {snsClient } from "../libs/snsClient.js";
 
 // Set the parameters
 const params = {
