+++ javascriptv3/example_code/sns/src/sns_publishtotopic.js
@@ -18,7 +18,7 @@ node sns_publishtotopic.js
 // snippet-start:[sns.JavaScript.topics.publishMessagesV3]
 // Import required AWS SDK clients and commands for Node.js
 import {PublishCommand } from "@aws-sdk/client-sns";
-import {snsClient } from "./libs/snsClient.js";
+import {snsClient } from "../libs/snsClient.js";
 
 // Set the parameters
 var params = {
