+++ javascriptv3/example_code/sns/src/sns_settopicattributes.js
@@ -20,7 +20,7 @@ node sns_settopicattributes.js
 
 // Import required AWS SDK clients and commands for Node.js
 import {SetTopicAttributesCommand } from "@aws-sdk/client-sns";
-import {snsClient } from "./libs/snsClient.js";
+import {snsClient } from "../libs/snsClient.js";
 
 // Set the parameters
 const params = {
