+++ javascriptv3/example_code/sns/src/sns_subscribeapp.js
@@ -18,7 +18,7 @@ node sns_subscribeapp.js
 // snippet-start:[sns.JavaScript.subscriptions.subscribeAppV3]
 // Import required AWS SDK clients and commands for Node.js
 import {SubscribeCommand } from "@aws-sdk/client-sns";
-import {snsClient } from "./libs/snsClient.js";
+import {snsClient } from "../libs/snsClient.js";
 
 // Set the parameters
 const params = {
