+++ javascriptv3/example_code/sns/src/sns_subscribelambda.js
@@ -19,7 +19,7 @@ node sns_subscribelambda.js
 // snippet-start:[sns.JavaScript.subscriptions.subscribeLambdaV3]
 // Import required AWS SDK clients and commands for Node.js
 import {SubscribeCommand } from "@aws-sdk/client-sns";
-import {snsClient } from "./libs/snsClient.js";
+import {snsClient } from "../libs/snsClient.js";
 
 // Set the parameters
 const params = {
